Fix integer to pointer of different size conversion warnings when using GCC for 32-bit platforms. The integer size in this case is hardcoded 64-bit while the pointer size is 32-bit.
Sponsored by:	Mellanox Technologies
MFC after:	2 weeks


Former-commit-id: c93fb1a712d000164ea1053035e30e0138cb53dd

+++ sys/ofed/drivers/infiniband/core/uverbs_main.c
@@ -262,7 +262,7 @@ parameter_list|,
 name|olen
 parameter_list|)
 define|\
-value|do {							\ 		(udata)->ops    =&uverbs_copy_ex;		\ 		(udata)->inbuf  = (void __user *)(ibuf);	\ 		(udata)->outbuf = (void __user *)(obuf);	\ 		(udata)->inlen  = (ilen);			\ 		(udata)->outlen = (olen);			\ 	} while (0)
+value|do {							\ 		(udata)->ops    =&uverbs_copy_ex;		\ 		(udata)->inbuf  = (void __user *)(unsigned long)(ibuf);	\ 		(udata)->outbuf = (void __user *)(unsigned long)(obuf);	\ 		(udata)->inlen  = (ilen);			\ 		(udata)->outlen = (olen);			\ 	} while (0)
 end_define
 
 begin_decl_stmt
